// src/interfaces/enemResult.ts
export interface EnemResult {
  id: string;
  inscription_number: string;
  name: string;
  cpf: string;
  foreign_language: string;
  languages_score: number;
  human_sciences_score: number;
  natural_sciences_score: number;
  math_score: number;
  essay_score: number;
  pdf_file: string;
  status: string;
  created_at: string;
  user_id: number;
}